refactor: move route path
