+++ apps/api/src/app/routes/trading/serializeQuoteAmountsAndCosts.ts
